chore: extract pending transaction logic to hook
